+++ src/main/java/com/project/banking/service/UserService.java
@@ -1,6 +1,5 @@
 package com.project.banking.service;
 
-import com.project.banking.model.Person;
 import com.project.banking.model.User;
 import com.project.banking.response.ApiResponse;
 import com.project.banking.response.UsersResponse;
@@ -13,9 +12,9 @@ import java.util.List;
 public interface UserService {
 
     ApiResponse createUser(User user) throws NoSuchAlgorithmException, InvalidKeySpecException;
-    ApiResponse updateUserPassword(Long id, User user) throws NoSuchAlgorithmException, InvalidKeySpecException;
-    ApiResponse deleteUser(Long id);
-    User getUserById(Long id);
+    ApiResponse updateUserPassword(Integer id, User user) throws NoSuchAlgorithmException, InvalidKeySpecException;
+    ApiResponse deleteUser(Integer id);
+    User getUserById(Integer id);
     public User getUserByUserName(String userName);
     List<User> getAllUsers();
     UsersResponse getUsersData() throws SQLException, ClassNotFoundException;
